+++ internal/compiler/generator/cpp.rs
@@ -3097,6 +3097,19 @@ fn compile_builtin_function_call(
             let window = access_window_field(ctx);
             format!("{window}.close_popup()")
         }
+        BuiltinFunction::SetSelectionOffsets => {
+            if let [llr::Expression::PropertyReference(pr), from, to] = arguments {
+                let item = access_member(pr, ctx);
+                let item_rc = access_item_rc(pr, ctx);
+                let window = access_window_field(ctx);
+                let start = compile_expression(from, ctx);
+                let end = compile_expression(to, ctx);
+
+                format!("slint_textinput_set_selection_offsets(&{item}, &{window}.handle(), &{item_rc}, static_cast<int>({start}), static_cast<int>({end}))")
+            } else {
+                panic!("internal error: invalid args to set-selection-offsets {:?}", arguments)
+            }
+        }
         BuiltinFunction::ItemMemberFunction(name) => {
             if let [llr::Expression::PropertyReference(pr)] = arguments {
                 let item = access_member(pr, ctx);